+++ packages/analytics/src/index.ts
@@ -18,3 +18,4 @@ export * from './tvm';
 export * from './dcf';
 export * from './relativeValue';
 export * from './capm';
+export * from './insider';